+++ scheduling/src/main/java/com/nabla/sdk/scheduling/scene/SchedulingBaseFragment.kt
@@ -17,4 +17,4 @@ public abstract class SchedulingBaseFragment : Fragment {
 }
 
 internal fun LayoutInflater.cloneWithNablaOverlay(): LayoutInflater =
-    cloneInContext(context?.withNablaVideoCallThemeOverlays())
+    cloneInContext(context?.withNablaSchedulingThemeOverlays())